DSH project 1 - Updated R code for data download
Originally we downloaded the data in CSV format, however that only took one section of the "description" field on which we are doing NLP clustering. This downloads the data to include the entire body of text in the IATI record as an extra column at the end of the dataframe. 

This code will not work on AH internet, it will work on gov wifi. This is due to security in dfid.