Update country names in list of stopwords
The stopwords list originally included country names and codes.  This was introduced when we attempted to cluster the data by theme (eg "Health") rather than country. As we are not clustering anymore, we now include country names and codes to give more context to a search. In this commit the names of each country, and their two and three letter country codes, are all removed from the stopwords list, this was from line 180 to line 920.