Fix python imports when running python code as scripts (#86)
* fix python imports

When we run scripts as a module, we use the module import
e.g. 'import from ips_python.constants'

but when we want to run a file as a script, we must reference the file
without the reference. e.g. 'import constants'

It is possible that this is not the best possible way to arrange our
imports, but it seems to work for now. Any input on this is welcome

- Also added tests to make sure that we pick up these errors in future